// src/app/models/RepoEfecActiTomaEst.ts
export class RepoEfecActiTomaEst {
	v_idcarga: string;
	ciclo: string;
	v_descofic: string; 
    v_nombreempr: string;
	periodo: string; 
	carga_entrega:  number;
	lecturas:  number;
	incidencias:  number;
	supervisadas:  number;
	porc_lecturas:  number;
	porc_incidencias:  number;
    porc_supervisadas:  number;
	con_dispositivo_movil:  number;
	con_hoja_ruta:  number;
  }
  